requirements.txt
roboflow
python-dotenv
ultralytics
opencv-python
numpy
supervision